fix(apps/playground): harden presence update bounds and liveness
Split presence-room tests into focused modules with shared helpers.
Enforce UTF-8 byte limits on raw update payloads before allowlisting,
publish normalized updates, reject future lastSeenAt values, and use
strict greater-than assertion for stale-update liveness refresh.
